windows vista

+++ src/Engine/Animation.cpp
@@ -15,7 +15,7 @@ SkeletalAnimation::SkeletalAnimation(const std::string& animationPath, Model* mo
 	currTime = 0.0;
 
 	//note: reserve(100)
-	finalBoneMatrices.reserve(animation->mDuration);
+	finalBoneMatrices.reserve(static_cast<const unsigned __int64>((animation->mDuration)));
 
 	for (int i = 0; i < animation->mDuration; i++)
 		finalBoneMatrices.push_back(glm::mat4(1.0f));
